Several Refractors - Desc
Using a cvzone wrapper on mediapipe - saves me time on optimization

Using a dict to store gestures for actions

Cvzone provides hand type, will use to make hand depend actions (such as clicking)

Still need to decide on which landmark on hand to use for mouse position

As well as normailzing (scaling the image coords to screen coords) properly. This including a scaling factor to assist with reaching the edge of the screen.